fix: split work-type editable tests to satisfy lint line limits
CI Build/Lint failed on noExcessiveLinesPerFunction in editable validation
and values test modules; split contract, duplicate-default, and default-
handling cases into sibling test files without changing behavior.
